+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Users Service, Demo user
 - Update frontend to store and send JWT tokens + auth api
+- Unified gateway - auth + fetch trades
 
 ## [0.9.019-c] - 2025-06-24
 ### Added